+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - move to edition 2024, MSRV 1.88
+- `CArray`, `CStringArray` and `CRange` have moved to the new `ffi-convert-extra-ctypes` crate. The core `ffi-convert` crate now only contains the conversion traits.
 
 ### Fixed
-- memory leak on array convertion error and performance / correctness improvement on array convertion
+- memory leak on array conversion error and performance / correctness improvement on array conversion
 
 ## [0.7.0] - 2026-04-13
 ### Added
