+++ Cargo.toml
@@ -3,5 +3,6 @@ resolver = "3"
 members = [
     "ffi-convert",
     "ffi-convert-derive",
+    "ffi-convert-extra-ctypes",
     "ffi-convert-tests",
 ]
