+++ ffi-convert-derive/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ffi-convert-derive"
 version = "0.8.0-pre"
-authors = ["Sonos"]
+authors = ["Sonos", "Anthony Reinette", "Thibaut Lorrain"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
 description = "Macros implementations of CReprOf, AsRust, CDrop traits from ffi-convert"
